Factor sftp connection & progress bar into separate class/module
