+++ app/importers/star_importer.rb
@@ -1,5 +1,7 @@
 module StarImporter
   include Importer
+  include ProgressBar
+
   # Any class using StarImporter should implement two methods:
   # export_file_name => string pointing to the name of the remote file to parse
   # header_converters => function that describes how to convert the headers on the remote files
@@ -12,27 +14,14 @@ module StarImporter
     @summer_school_local_ids = options[:summer_school_local_ids]    # For importing only summer school students
   end
 
-  def sftp_info_present?
-    ENV['STAR_SFTP_HOST'].present? &&
-    ENV['STAR_SFTP_USER'].present? &&
-    ENV['STAR_SFTP_PASSWORD'].present?
-  end
-
   def connect_and_import
-    if sftp_info_present?
-      Net::SFTP.start(
-        ENV['STAR_SFTP_HOST'],
-        ENV['STAR_SFTP_USER'],
-        password: ENV['STAR_SFTP_PASSWORD']
-        ) do |sftp|
-        # For documentation on Net::SFTP::Operations::File, see
-        # http://net-ssh.github.io/sftp/v2/api/classes/Net/SFTP/Operations/File.html
-        file = sftp.download!(export_file_name)
-        import(file)
-      end
-    else
-      raise "SFTP information missing"
-    end
+    sftp = SftpClient.new({
+      user: ENV['STAR_SFTP_USER'],
+      host: ENV['STAR_SFTP_HOST'],
+      password: ENV['STAR_SFTP_PASSWORD']
+    }).start
+    file = sftp.download!(export_file_name)
+    import(file)
   end
 
   def import(file)
